+++ src/test/kotlin/com/rickbusarow/ktrules/EditorConfigPropertiesTest.kt
@@ -46,6 +46,7 @@ class EditorConfigPropertiesTest : Tests {
       [{*.kt,*.kts}]
       ktlint_kt-rules_kdoc-indent-after-leading-asterisk = enabled
       ktlint_kt-rules_kdoc-leading-asterisk = enabled
+      ktlint_kt-rules_kdoc-tag-order = enabled
       ktlint_kt-rules_kdoc-wrapping = enabled
       ktlint_kt-rules_no-duplicate-copyright-header = enabled
       ktlint_kt-rules_no-leading-blank-lines = enabled